Globalize translator and remove DEFAULT slot.
In `lookup-key', translations should have lower priority than the keymap
arguments.  See new test for an example where we want translations to be iterated
after we've checked all the keymaps.

Default argument is not useful if we can check matches with the second return
value.  In the end, it's up to the caller to decide what to do when nothing matches.
It's unclear whether we want the behaviour to change locally with regard to the keymap.
Should we want keymap-local defaults, we would have to accommodate `lookup-key'
since it can take multiple keymaps as argument: which default do we pick?
Since this is a rather klunky design, we leave it out for now.